telegram_bot/requirements.txt
torch
torchvision
pillow
python-telegram-bot